The big squash:
This is the "initial" commit for Reportobello, a Free and Open Source PDF
and report generation API based on Typst. I've been working on this for a
while, and am deciding to release it now so that I can get some initial
feedback on it. I've done a lot of work, rework, and redesigns, so I'm
condensing all the messy history into one snapshot and going from there.

Why did I build Reportobello? Because everytime I want to integrate PDF reports
for a client or project I have to use a different API or library. What I want
is one API/library for all my reporting needs, and have SDK support for all the
programming languages I'm integrating with. When I heard about Typst, I saw
that I could have a powerful, configurable, and fast templating language for
building reports. Since Typst is primarily a CLI tool, I decided it would be
nice to have an API wrapper around it for building reports. In the future I
might add support for other templating languages, but Typst checks all the
boxes for me right now.

While Typst is very powerful, it does take some getting used to. To make
using Reportobello easier, I also created `x2typst`, a tool for converting
PDFs, DOC/DOCX, and Markdown files to Typst. It can't easily be used as a
stand-alone tool, but you can use it by converting your files on the home page
of https://reportobello.com (or on the home page after logging in).