Add util file for features not (yet) in Typst:
There are certain features I want to use within Typst which are not yet
supported. By adding a `.typ` util file into the build directory for each
template users can import extra functions, globals, etc. that are useful
in making reports, such as date parsing. The mechanism for doing this right
now is not great since it assumes you have the `reportobello.typ` file in
your directory, which ideally would be a package that the user can install.
It isn't that simple though, but could be made easier in the future.