Convert `.typ` file to package:
This improves upon the last commit by allowing users to indirectly import the
`data` global variable without having to pass the `data.json` boilerplate. This
also gives Reportobello more control over how it injects the JSON data into the
template, allowing users to use different data files (ie, for the CLI).

The issue now is making this work for the CLI. The code is backwards compatible
meaning you can still use the `json("data.json")` method if you like, but for
the code to work witn the CLI, the package needs to be downloaded and installed
locally on the user's computer. This shouldn't be too hard, you just need to
copy the files to the user's computer. The issue though is that code is in
another repo, so I will probably need to add some script to download and pull
all that in.