Make uploaded file contents read only:
Because every uploaded file is symlinked to the same underlying file, if there
ever is a situation where the file is overridden (either on accident or
maliciously), there would be no way to prevent it. Now the file is marked as
read-only, which should fix the issue of the files getting overridden.