Fix post-login issues with `SameSite=strict` cookies:
Turns out that `strict` cookies sometimes have issues after logging via GitHub
because the browser still things the site has been loaded from a 3rd party (at
least that is what I think is happening). Even after reloading and force
reloading the site does send the right cookies, only after you navigate to
another page and go back does it work. The solution is to add a query param to
the `/` URL to tell the `index.html` page to reload if it sees it, allowing for
the correct page to load. This is a real pain and shouldn't be necessary, but
alas.