Fix `uv` not installing packages correctly:
There is an issue while installing pymupdfpro, which isn't needed by the
server directly, but was being specified in the dev dependencies. This package
is installable with normal pip, but is causing issues with uv. I can't seem
to find a relevant issue for this, but from what I can tell, uv is having
issues installing `manylinux` wheels on the `linux` platform. For now
everything is "working", but hopefully that issue is fixed.