Use `Host` header to determine redirect location:
Depending on the browser/server setup, the domain env var might not match what
the server sends it (for example, using Reportobello behind a reverse proxy or
VM network mapping). Essentially, if the caller is making a request to a
specific `Host`, it makes sense we re-direct them back to the same `Host`.